refactor: 将 Agent 相关组件统一归集至 agents 模块
- 将 AgentModalContent 从 SettingsModal/contents 迁至 agents/components
- 将 AssistantEditDrawer 从 guid/components 迁至 agents/components
- 使 agents 模块自包含，降低跨模块耦合
